embrace pytest structure
