+++ tests/ulyanov_et_al_2016/test_loss.py
@@ -1,26 +1,7 @@
-import pytest
-
 from torch.nn.functional import mse_loss
 
 from pystiche.image import extract_batch_size
-from pystiche_papers import ulyanov_et_al_2016 as paper
 from pystiche_papers.ulyanov_et_al_2016 import loss
-
-from .asserts import assert_image_downloads_correctly, assert_image_is_downloadable
-
-
-def test_ulyanov_et_al_2016_images_smoke(subtests):
-    for name, image in paper.ulyanov_et_al_2016_images():
-        with subtests.test(name=name):
-            assert_image_is_downloadable(image)
-
-
-@pytest.mark.large_download
-@pytest.mark.slow
-def test_ulyanov_et_al_2016_images(subtests):
-    for name, image in paper.ulyanov_et_al_2016_images():
-        with subtests.test(name=name):
-            assert_image_downloads_correctly(image)
 
 
 def test_ulyanov_et_al_2016_content_loss(
